requirements.txt
pandas
openai
tqdm
python-dotenv